+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Stories
 
 ### Bug fixes
+* [UIREC-62](https://issues.folio.org/browse/UIREC-62) Hotlink POL number in receiving to POL in orders
 * [UIREC-61](https://issues.folio.org/browse/UIREC-61) Caption not displayed when receiving or undeceiving pieces
 
 ## [1.0.0](https://github.com/folio-org/ui-receiving/tree/v1.0.0) (2020-03-13)